[ChessRouler] Move he chessColor utils into its own file
