+++ util-kit/src/lib/ChessRouler/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './chessColor';
+export * from './misc';
